Move PolynomialBases into separate header files
